requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
altair==4.2.2
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
biopython==1.81
bleach==6.0.0
blinker==1.6.2
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
confection==0.0.4
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
cymem==2.0.7
Cython==0.29.14
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
flatbuffers==23.5.8
fonttools==4.39.3
fqdn==1.5.1
gast==0.4.0
gensim==3.8.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
kiwisolver==1.4.4
langcodes==3.3.0
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
multiprocess==0.70.14
murmurhash==1.0.9
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook_shim==0.2.2
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.0
pathy==0.10.1
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
plac==1.1.3
platformdirs==3.2.0
plotly==5.14.1
pluggy==1.0.0
pox==0.3.2
ppft==1.7.6.6
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.5
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
PyExecJS==1.5.1
pygame==2.3.0
Pygments==2.15.1
pymatch==0.3.4
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
quantiprot==0.2.4
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.4
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
statsmodels==0.13.5
streamlit==1.21.0
streamlit-option-menu==0.3.2
sympy==1.11.1
tenacity==8.2.2
termcolor==2.3.0
terminado==0.17.1
thinc==8.1.10
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
Unidecode==1.3.6
uri-template==1.2.0
urllib3==1.26.15
validators==0.20.0
wasabi==0.10.1
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.4
widgetsnbextension==4.0.7
wordcloud==1.9.1.1
wrapt==1.15.0
y-py==0.5.9
ypy-websocket==0.8.2
zipp==3.15.0
